% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{ad_hoc_blankmodern_ifequals}
\alias{ad_hoc_blankmodern_ifequals}
\title{Blank out modern use if modern cp and any CP are equal}
\usage{
ad_hoc_blankmodern_ifequals(obs)
}
\arguments{
\item{obs}{data frame with contraceptive_use_any, contraceptive_use_modern, contraceptive_use_traditional}
}
\value{
data frame with contraceptive_use_modern set to NA if it equals contraceptive_use_any and contraceptive_use_traditional is NA
}
\description{
Blank out modern use if modern cp and any CP are equal
}
\keyword{internal}
