% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{ad_hoc_calculate_cp_trad}
\alias{ad_hoc_calculate_cp_trad}
\title{Ad hoc calculation of CP Traditional}
\usage{
ad_hoc_calculate_cp_trad(obs)
}
\arguments{
\item{obs}{data frame with contraceptive_use_any, contraceptive_use_modern, contraceptive_use_traditional}
}
\value{
data frame with contraceptive_use_traditional filled in where possible
}
\description{
Taken from UNPD/fpemlocal: try to calculate cp traditional from any and modern
}
\keyword{internal}
