% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bias_info.R
\name{add_bias_info}
\alias{add_bias_info}
\title{Add bias info}
\usage{
add_bias_info(data, married)
}
\arguments{
\item{data}{survey data}

\item{married}{boolean}
}
\value{
tibble based on data with added columns for any_bias, trad_bias, tradorany_bias
}
\description{
Add columns for any_bias, trad_bias, tradorany_bias
}
\keyword{internal}
