% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_natdata_for_aggregates.R
\name{add_natdata_for_aggregates}
\alias{add_natdata_for_aggregates}
\title{Add national data for aggregates}
\usage{
add_natdata_for_aggregates(national_dat_df, regions_dat, is_married)
}
\arguments{
\item{national_dat_df}{national survey data prior to processing}

\item{regions_dat}{region metadata}

\item{is_married}{boolean}
}
\value{
Processed national data with region_code set to "National"
}
\description{
This function processes national data and adds it to the aggregate dataset.
}
\keyword{internal}
