% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outliers.R
\name{add_outlier_related_columns}
\alias{add_outlier_related_columns}
\title{Add possible-outlier related columns}
\usage{
add_outlier_related_columns(dat)
}
\arguments{
\item{dat}{survey data}
}
\value{
data frame with new columns possible_outlier (1 = possible outlier, 0 = not a possible outlier) and possible_outlier_userinput (NA)
}
\description{
To survey data, add columns possible_outliers and possible_outlier_userinput. Possible outliers are assessed
using the `assign_outliers` function.
}
\keyword{internal}
