% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uncertainty_in_obs.R
\name{add_uncertainty_in_obs}
\alias{add_uncertainty_in_obs}
\title{Calculate uncertainty in observations for fit to one marital group}
\usage{
add_uncertainty_in_obs(
  fit,
  samples_scale = NULL,
  perc_low = 0.025,
  perc_up = 0.975,
  add_nse_uncertainty = TRUE
)
}
\arguments{
\item{fit}{list with fit information, including data and posterior samples}

\item{samples_scale}{if provided, used instead of fit$samples$draws("DMX_scale")}

\item{perc_low}{percentile for lower bound of uncertainty interval, defaults to 0.025}

\item{perc_up}{percentile for upper bound of uncertainty interval, defaults to 0.975}

\item{add_nse_uncertainty}{boolean, if TRUE (default), adds NSE uncertainty in addition to sampling error uncertainty}

\item{is_married}{boolean}
}
\value{
survey data with columns for point estimates and uncertainty intervals added
}
\description{
For one marital group, adds uncertainty in observations based on sampling errors and, if specified, also adds NSE uncertainty.
To be used for global fits for one marital group.
}
