% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uncertainty_in_obs.R
\name{add_uncertainty_in_obs_allwomen}
\alias{add_uncertainty_in_obs_allwomen}
\title{Calculate uncertainty in observations for all-women fit}
\usage{
add_uncertainty_in_obs_allwomen(
  samples,
  data_allwomen,
  is_married = TRUE,
  perc_low = 0.025,
  perc_up = 0.975,
  add_nse_uncertainty = TRUE
)
}
\arguments{
\item{samples}{posterior::draws object from cmdstanr fit}

\item{data_allwomen}{list with data for all women used in model fitting}

\item{is_married}{boolean}

\item{perc_low}{percentile for lower bound of uncertainty interval, defaults to 0.025}

\item{perc_up}{percentile for upper bound of uncertainty interval, defaults to 0.975}

\item{add_nse_uncertainty}{boolean, if TRUE (default), adds NSE uncertainty in addition to sampling error uncertainty}
}
\value{
survey data for both marital groups with columns for point estimates and uncertainty intervals added
}
\description{
For all-women FPET fit, adds uncertainty in observations based on sampling errors and, if specified, also adds NSE uncertainty.
}
\keyword{internal}
