% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outliers.R
\name{assign_outliers}
\alias{assign_outliers}
\title{Assign outliers}
\usage{
assign_outliers(fp_dat, outlier_record_ids = NULL, married)
}
\arguments{
\item{fp_dat}{survey data}

\item{outlier_record_ids}{record_ids of outliers (if any)}

\item{married}{boolean, needed when bias needs to be assessed}
}
\value{
data frame with new column nooutlier (1 = not an outlier, 0 = outlier)
}
\description{
Function to assign outliers based on a set of rules.
If columns possible_outlier or possible_outlier_userinput are present, these are used to define outliers.
}
\keyword{internal}
