% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{Compile a Stan model}
\usage{
compile_model(
  variational,
  nthreads_variational,
  force_recompile,
  stan_file_path
)
}
\arguments{
\item{variational}{boolean}

\item{nthreads_variational}{if variational, number of threads to use}

\item{force_recompile}{boolean}

\item{stan_file_path}{path to the stan file}
}
\value{
cmdstanr::cmdstan_model object
}
\description{
Compile a Stan model
}
\keyword{internal}
