% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/construct_popweightmatrix_for_oneaggregateunit.R
\name{construct_popweightmatrix_for_oneaggregateunit}
\alias{construct_popweightmatrix_for_oneaggregateunit}
\title{Construct population weight matrix for one aggregate unit}
\usage{
construct_popweightmatrix_for_oneaggregateunit(
  aggregation_meta_one_unit,
  population_selectedyear,
  name_region_index = "region_index"
)
}
\arguments{
\item{aggregation_meta_one_unit}{tibble with columns agg_unit_name and region_index}

\item{population_selectedyear}{tibble with region_index and population, needs to include all regions}
}
\value{
row vector with weights for each region in the aggregate unit (0 if region is not included)
}
\description{
Constructs a row vector of population weights for regions in a single aggregate unit.
}
\keyword{#}
\keyword{example_creatingaggregates.R}
\keyword{in}
\keyword{internal}
\keyword{now}
\keyword{repo}
\keyword{right}
\keyword{submeta}
