% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emu_global_hyperparameters.R
\docType{data}
\name{emu_global_hyperparameters}
\alias{emu_global_hyperparameters}
\title{EMU hyperparameters}
\format{
## `emu_global_hyperparameters`
tibble with
\describe{
  \item{ss_type}{vistis, clients, facilities, users}
  \item{mean_log_sigma_type}{mean log SD for type}
  \item{hierarchical sigma}{across-population SD for type}
}
}
\source{
global fit 2025
}
\usage{
emu_global_hyperparameters
}
\description{
EMU hyperparameters
}
\keyword{internal}
