% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_fit_objects.R
\docType{data}
\name{fit3_married}
\alias{fit3_married}
\title{Global fit object fit3_married}
\format{
## `fit3_married`
list with global fit objects including
\describe{
  \item{post_summ}{data frame with posterior summaries for relevant parameters}
  \item{geo_unit_index}{data frame with geo_unit info}
  \item{source_index}{data frame with sources of survey data with indices used in the model}
  \item{time_index}{data frame with geo_unit info}
  \item{other settings}{spline settings, t_star, etc}
}
}
\source{
global fit 2025
}
\usage{
fit3_married
}
\description{
Global fit object fit3_married
}
\keyword{internal}
