% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fpem.R
\name{fit_fpem}
\alias{fit_fpem}
\title{Fit FPET national or subnational local models for ALL women, with or without EMUs.}
\usage{
fit_fpem(
  survey_df,
  population_df,
  service_statistic_df = NULL,
  subnational = FALSE,
  national_dat_df = NULL,
  chains = 4,
  iter_sampling = 200,
  iter_warmup = 150,
  seed = 1234,
  regions_dat,
  end_year = 2030
)
}
\arguments{
\item{survey_df}{Survey data for model fitting, filtered to pop of interest}

\item{population_df}{Population data, filtered to pop of interest}

\item{service_statistic_df}{EMU data, filtered to pop of interest}

\item{subnational}{logical, TRUE if fitting to subnational data.
For subnational fits, survey_df may contain all or just 1 region.}

\item{national_dat_df}{National data to be used/displayed for subnational runs,
filtered to pop of interest. Model is fitted to surveys in national_dat_df that are not in survey_df.}

\item{chains}{Number of chains for Stan}

\item{iter_sampling}{Number of iterations for sampling}

\item{iter_warmup}{Number of iterations for warmup}

\item{seed}{Random seed for Stan}

\item{regions_dat}{Data frame with region information}

\item{end_year}{End year for the model, default is 2030}
}
\value{
results list with estimates, samples, data_allwomen, observations,
 and dat_emu = dat_emu
}
\description{
Fit FPET national or subnational local models for ALL women, with or without EMUs.
}
