% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_allwomen.R
\name{fit_model_allwomen}
\alias{fit_model_allwomen}
\title{Fit model to produce all women estimates}
\usage{
fit_model_allwomen(
  data_allwomen,
  runstep = "local_national",
  population_df,
  emu_list = NULL,
  end_year,
  chains = 4,
  iter_sampling = 200,
  iter_warmup = 150,
  add_inits = FALSE,
  add_sample = TRUE,
  compile_model = TRUE,
  force_recompile = FALSE,
  seed = 1234,
  refresh = 50,
  adapt_delta = 0.9,
  max_treedepth = 14,
  variational = FALSE,
  nthreads_variational = 8
)
}
\arguments{
\item{data_allwomen}{list with input data for married and data for unmarried women}

\item{runstep}{local_national or local_subnational}

\item{population_df}{pop counts for each geounit-year combi}

\item{emu_list}{emu info from `service_statistics_getstandata`}

\item{chains}{Number of chains for Stan}

\item{iter_sampling}{Number of iterations for sampling}

\item{iter_warmup}{Number of iterations for warmup}

\item{add_inits}{Use function to sample initial values? Defaults to FALSE}

\item{add_sample}{TRUE}

\item{compile_model}{TRUE}

\item{force_recompile}{FALSE}

\item{seed}{Random seed for Stan}

\item{refresh}{10}

\item{adapt_delta}{0.9}

\item{max_treedepth}{14}

\item{variational}{FALSE}

\item{nthreads_variational}{8}
}
\value{
results list with stan_model, stan_data, samples, runstep
}
\description{
This function is called from fit_fpem,
to do the model fitting for all women
}
\keyword{internal}
