% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet2-package.R
\docType{package}
\name{fpet2-package}
\alias{fpet2}
\alias{fpet2-package}
\title{fpet2: An R Package for Estimation and Forecasting of Family Planning Indicators}
\description{
The fpet2 package implements Bayesian models for estimation and forecasting of family planning Indicators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlkemaLab/fpet2}
  \item \url{https://alkemalab.github.io/fpet2025/}
  \item \url{https://alkemalab.github.io/fpet2/}
  \item Report bugs at \url{https://github.com/AlkemaLab/fpet2/issues}
}

}
\author{
\strong{Maintainer}: Leontine Alkema \email{lalkema@umass.edu}

Authors:
\itemize{
  \item Herbert Susmann \email{herbps10@gmail.com}
  \item Shauna Mooney
  \item Evan Ray
}

}
\keyword{internal}
