% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{get_effective_samplesize}
\alias{get_effective_samplesize}
\title{Get effective sample size from proportion and its standard error}
\usage{
get_effective_samplesize(p, se_p)
}
\arguments{
\item{p}{proportion}

\item{se_p}{standard error of the proportion}
}
\value{
effective sample size
}
\description{
Get effective sample size from proportion and its standard error
}
\keyword{internal}
