% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates_globalruns}
\alias{get_estimates_globalruns}
\title{Get estimates from global runs}
\usage{
get_estimates_globalruns(
  samples,
  geo_unit_index,
  time_index,
  marital_status = "married",
  return_samples = FALSE,
  add_trad = TRUE
)
}
\arguments{
\item{samples}{tibble of posterior samples with columns for indicators, geo_unit_index, time_index, and value}

\item{geo_unit_index}{geo_unit_index from model fit}

\item{time_index}{time_index from model fit}

\item{marital_status}{string, "married" or "unmarried"}

\item{return_samples}{boolean, if TRUE, returns posterior samples along with summaries}

\item{add_trad}{boolean, if TRUE, includes traditional use in the output}
}
\value{
tibble with estimates (mean and percentiles) for each indicator and marital status group
}
\description{
Let's keep this function as we use it for the global runs
it worked for all women but not yet aggregates
}
