% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_transition.R
\name{get_hier_data_transition}
\alias{get_hier_data_transition}
\title{Get hierarchical data for transition model}
\usage{
get_hier_data_transition(
  geo_unit_index,
  prefix = "",
  stan_data_settings,
  stan_spline_data,
  hierarchical_terms_and_fixed,
  global_fit = NULL
)
}
\arguments{
\item{geo_unit_index}{geo_unit_index for model fit}

\item{prefix}{prefix to add to parameter names, e.g., "d_" or "z_"}

\item{stan_data_settings}{list of stan data settings, e.g., bounds and priors}

\item{stan_spline_data}{stan data for splines, needed for dimension of splines coeff vector}

\item{hierarchical_terms_and_fixed}{list with info on levels and what's fixed from other function}

\item{global_fit}{fit object from a global fit, or NULL if not doing a local fit}
}
\value{
list with hier_stan_data_settings, hier_stan_data, and hier_data
}
\description{
Get hierarchical data for transition model
}
\keyword{internal}
