% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_data_transition.R
\name{get_hier_data_transition_justomega}
\alias{get_hier_data_transition_justomega}
\title{Get hierarchical data for Omegas in transition model}
\usage{
get_hier_data_transition_justomega(
  geo_unit_index,
  prefix = "",
  stan_data_settings,
  hierarchical_terms_and_fixed,
  global_fit = NULL,
  runstep
)
}
\arguments{
\item{geo_unit_index}{geo_unit_index for model fit}

\item{prefix}{prefix to add to parameter names, e.g., "d_" or "z_"}

\item{stan_data_settings}{list of stan data settings, e.g., bounds and priors}

\item{hierarchical_terms_and_fixed}{list with info on levels and what's fixed from other function}

\item{global_fit}{fit object from a global fit, or NULL if not doing a local fit}

\item{runstep}{what model is fitted}
}
\value{
list with hier_stan_data_settings, hier_stan_data, and hier_data
}
\description{
Perhaps to combine with `get_hier_data_transition`
}
\keyword{internal}
