% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levelinfo_fitmodel.R
\name{get_hierlevels_omegatrad}
\alias{get_hierlevels_omegatrad}
\title{Get list with information on hierarchical levels and which ones to fix for omegatrad}
\usage{
get_hierlevels_omegatrad(
  runstep,
  hierarchical_terms = NULL,
  global_fit = NULL,
  prefix = "",
  add_subnational_hierarchy = NULL
)
}
\arguments{
\item{runstep}{character, from model fit}

\item{hierarchical_terms}{list of hierarchical terms from fit_model inputs or NULL}

\item{global_fit}{fit object from a global fit}

\item{prefix}{character, prefix for process model (should be "z_" for traditional)}

\item{add_subnational_hierarchy}{character, name of subnational hierarchy level to add, e.g., "subnat"}
}
\value{
list with hierarchical terms and what's fixed
}
\description{
Get list with information on hierarchical levels and which ones to fix for omegatrad
}
\keyword{internal}
