% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_modern_fromdemandanddemandsat}
\alias{get_modern_fromdemandanddemandsat}
\title{Get modern from demand and demand satisfied}
\usage{
get_modern_fromdemandanddemandsat(demand, demand_satisfied)
}
\arguments{
\item{demand}{demand}

\item{demand_satisfied}{demand satisfied (modern/demand)}
}
\value{
modern
}
\description{
Model returns Eta (demandsat) and d_Eta (demand)
# real demand = inv_tr_eta(tr_d_Eta_obs[geo_unit[i], time[i]]);
# modern[i] = inv_tr_eta(tr_Eta_obs[geo_unit[i], time[i]]) * demand;
# logit_unmetovernonmodern[i] = logit((demand - modern[i])/(1-modern[i]));
}
\keyword{internal}
