% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nat_for_fitting.R
\name{get_nat_for_fitting}
\alias{get_nat_for_fitting}
\title{Get national data for subnational fitting}
\usage{
get_nat_for_fitting(national_dat_df, survey_df_use)
}
\arguments{
\item{national_dat_df}{National survey data, for specific marital group and div code}

\item{survey_df_use}{Subnational survey data used in fitting, for specific marital group and div code}
}
\value{
Dataframe with national data if not present for all regions in subnational survey data
}
\description{
Get national data for subnational fitting
}
\keyword{internal}
