% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonse_standata.R
\name{get_nonse_standata}
\alias{get_nonse_standata}
\title{Get stan data for non-sampling errors (nonse)}
\usage{
get_nonse_standata(fix_nonse, global_fit, source_index, prefix = "DM1_")
}
\arguments{
\item{fix_nonse}{logical, whether to fix nonse parameters to values from a global fit}

\item{global_fit}{fit object from a global fit, if fix_nonse is TRUE}

\item{source_index}{source_index data frame for the local fit}

\item{prefix}{prefix for nonse parameters, defaults to "DM1_"}
}
\value{
list of stan data elements for nonse
}
\description{
Get stan data for non-sampling errors (nonse)
}
\keyword{internal}
