% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dir_related.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get directory of output, located in parent of wd}
\usage{
get_output_dir(...)
}
\arguments{
\item{...}{character string subdirectories of the bayestransition_output
folder}
}
\value{
character string of the full path to the output directory
}
\description{
Get directory of output, located in parent of wd
}
\keyword{internal}
