% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_summaries_andfindpar.R
\name{get_posterior_summaries_andfindpar}
\alias{get_posterior_summaries_andfindpar}
\title{Get posterior summaries for parameters to fix}
\usage{
get_posterior_summaries_andfindpar(
  fit,
  process_indicator_prefixes = c("", "d_", "z_"),
  dm_indicator_prefixes = c("DM1_", "DM2_", "DM3_"),
  process_params = c("Betas_raw", "Betas_sigma", "Ptilde_raw", "Ptilde_sigma",
    "Omega_raw", "Omega_sigma"),
  process_smoothing_params = c("Rho", "Tau"),
  process_subnat_smoothing_params = c("rho_correlationeps"),
  dm_params = c("nonse", "sdbias", "rho_pma"),
  dm_outlier_params = c("global_shrinkage_dm", "caux_dm")
)
}
\arguments{
\item{fit}{fit object from a global fit}

\item{process_indicator_prefixes}{defaults to c("", "d_", "z_")}

\item{dm_indicator_prefixes}{defaults to c("DM1_", "DM2_", "DM3_")}

\item{process_params}{defaults to _raw and _sigma for Betas, Omega, and Ptilde}

\item{process_smoothing_params}{defaults to c("Rho", "Tau")}

\item{process_subnat_smoothing_params}{defaults to c("rho_correlationeps")}

\item{dm_params}{defaults to c("nonse", "sdbias", "rho_pma")}

\item{dm_outlier_params}{defaults to c("global_shrinkage_dm", "caux_dm")}
}
\value{
tibble with posterior summaries for parameters to fix
}
\description{
This function collects posterior summaries for a set of parameters
}
\keyword{internal}
