% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dir_related.R
\name{get_relative_output_dir}
\alias{get_relative_output_dir}
\title{Get relative output directory path}
\usage{
get_relative_output_dir(folder_name)
}
\arguments{
\item{folder_name}{Character. The name of the subfolder inside `bayestransition_output`.}
}
\value{
character string of the full path to the output directory
}
\description{
Constructs a relative path to a subfolder inside `bayestransition_output`.
}
\keyword{internal}
