% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Get residuals}
\usage{
get_residuals(fit, is_married)
}
\arguments{
\item{fit}{fit object from a global fit}

\item{is_married}{boolean}
}
\value{
tibble with survey data and residuals
}
\description{
Get residuals
}
\keyword{internal}
