% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{get_se}
\alias{get_se}
\title{Get Binomial standard error from proportion and sample size}
\usage{
get_se(p, n)
}
\arguments{
\item{p}{proportion}

\item{n}{sample size}
}
\value{
standard error
}
\description{
Get Binomial standard error from proportion and sample size
}
\keyword{internal}
