% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{get_se_logitprop}
\alias{get_se_logitprop}
\title{Get standard error on logit scale from proportion and its standard error}
\usage{
get_se_logitprop(prop, se_prop)
}
\arguments{
\item{prop}{proportion}

\item{se_prop}{standard error of the proportion (original scale)}
}
\value{
standard error on logit scale
}
\description{
Get standard error on logit scale from proportion and its standard error
}
\keyword{internal}
