% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{get_se_sequentialdm}
\alias{get_se_sequentialdm}
\title{Get standard errors for FP indicators in sequential DM approach}
\usage{
get_se_sequentialdm(nat_data, use_small_SE_setting = TRUE)
}
\arguments{
\item{nat_data}{data frame with survey data}

\item{use_small_SE_setting}{logical, if TRUE (default), use smaller effective sample sizes when imputing SEs}
}
\value{
tibble with SEs added
}
\description{
Get standard errors for FP indicators in sequential DM approach
}
\keyword{internal}
