% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoothing_standata.R
\name{get_smoothing_standata}
\alias{get_smoothing_standata}
\title{Get smoothing standata}
\usage{
get_smoothing_standata(fix_smoothing, global_fit, prefix = "")
}
\arguments{
\item{fix_smoothing}{logical, whether to fix smoothing parameters}

\item{global_fit}{fit object from a global fit, used to extract smoothing parameters if fix_smoothing is TRUE}

\item{prefix}{prefix to add to the parameter names, defaults to ""}
}
\value{
list with smoothing standata
}
\description{
Get smoothing standata
}
\keyword{internal}
