% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_spline_data.R
\name{get_stan_spline_data}
\alias{get_stan_spline_data}
\title{Get spline data inputs for Stan model}
\usage{
get_stan_spline_data(num_knots, spline_degree)
}
\arguments{
\item{num_knots}{number of knots}

\item{spline_degree}{degree of spline}
}
\value{
list of spline data inputs for Stan model
}
\description{
Get spline data inputs for Stan model
}
\keyword{internal}
