% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_unmet_fromdemandandmodern}
\alias{get_unmet_fromdemandandmodern}
\title{Get unmet from demand and modern}
\usage{
get_unmet_fromdemandandmodern(demand, modern)
}
\arguments{
\item{demand}{demand}

\item{modern}{modern}
}
\value{
unmet
}
\description{
Get unmet from demand and modern
}
\keyword{internal}
