% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{indicate_rounding}
\alias{indicate_rounding}
\title{Indicate rounding}
\usage{
indicate_rounding(x)
}
\arguments{
\item{x}{proportion to check}
}
\value{
indicator variable (1 if x < 0.001, else 0)
}
\description{
Taken from UNPD/fpemlocal
}
\keyword{internal}
