% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_fun}
\alias{init_fun}
\title{Initialization function for stan model}
\usage{
init_fun(chain_id, stan_data)
}
\arguments{
\item{chain_id}{integer chain id}

\item{stan_data}{list of stan data}
}
\value{
list of initial values for parameters
}
\description{
Initialization function for stan model
}
\keyword{internal}
