% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{input value}
}
\value{
inverse logit of x
}
\description{
Inverse logit function
}
\keyword{internal}
