% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dir_related.R
\name{load_fit}
\alias{load_fit}
\title{Load fit}
\usage{
load_fit(path)
}
\arguments{
\item{path}{path to the folder containing the model fit}
}
\value{
object of class fpemplus
}
\description{
Load a fit object. This fit should be saved in a folder that contains
a file named `fit.rds` with the fpemplus object as well as csv files produced
by cmdstanr as a result of the estimation process.
}
\keyword{internal}
