% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_se_sequentialdm.R
\name{logit}
\alias{logit}
\title{probit <- function(x) pnorm(x)
# stan
#phiapprox <- function(x) inv_logit(0.07056 * x ^ 3 + 1.5976 * x)
inv_probit <- function(x) qnorm(x)}
\usage{
logit(x)
}
\arguments{
\item{x}{input value}
}
\value{
logit of x
}
\description{
#' Get standard error on inverse probit scale from proportion and its standard error
#'
#' @param prop proportion
#' @param se_prop standard error of the proportion (original scale)
#'
#' @returns standard error on inverse probit scale
#' @keywords internal
get_se_invprobitprop <- function(prop, se_prop){
  # se_prop*abs(derivate evaluated at prop)
  se_prop*1/dnorm(inv_probit(prop))
}
Logit function
}
\keyword{internal}
