% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eps.R
\name{plot_eps}
\alias{plot_eps}
\title{Plot Epsilon Innovations}
\usage{
plot_eps(fit, is_married, estimates_df, add_trad = FALSE)
}
\arguments{
\item{fit}{fit object from a global fit}

\item{is_married}{boolean}

\item{estimates_df}{dataframe of estimates}

\item{add_trad}{boolean, whether to add traditional epsilon innovations}
}
\value{
NULL, saves plots to fit$output_dir/eps.pdf
}
\description{
Plot Epsilon Innovations
}
\keyword{internal}
