% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates_local.R
\name{plot_estimates_local}
\alias{plot_estimates_local}
\title{Plot FP estimates for one country/region, marital status, and indicator}
\usage{
plot_estimates_local(
  estimates,
  filtered_data,
  indicator_select,
  estimates2 = NULL,
  estimates3 = NULL,
  estimates4 = NULL,
  modelnames = c("model1", "model2"),
  cols_sourcetypes = c(DHS = "red", DHS0 = "red", MICS = "blue", PMA = "darkgreen", Other
    = "orange", `National survey` = "purple")
)
}
\arguments{
\item{estimates}{estimates}

\item{filtered_data}{observations}

\item{indicator_select}{indicator}

\item{estimates2}{optional second model estimates to compare}

\item{estimates3}{optional third model estimates to compare}

\item{estimates4}{optional fourth model estimates to compare}

\item{modelnames}{names of models to use in legend if comparing more than one model}

\item{cols_sourcetypes}{colors to use for data source types}
}
\value{
ggplot object
}
\description{
Plot FP estimates for one country/region, marital status, and indicator
}
\keyword{internal}
