% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hierchecks.R
\name{plot_hierchecks}
\alias{plot_hierchecks}
\title{Plot hierarchical checks}
\usage{
plot_hierchecks(
  fit,
  is_married,
  add_priors_transitions = TRUE,
  add_priors_trad = FALSE,
  add_trad = FALSE
)
}
\arguments{
\item{fit}{fit object from a global fit}

\item{is_married}{logical, is this a married model?}

\item{add_priors_transitions}{Plot prior-posts for sigmas?}

\item{add_priors_trad}{Plot prior-posts for trad sigmas?}

\item{add_trad}{Plot trad parameters?}
}
\value{
NULL, but saves a pdf with plots to the output directory
}
\description{
Plot hierarchical checks
}
\keyword{internal}
