% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot residuals}
\usage{
plot_residuals(fit, is_married)
}
\arguments{
\item{fit}{fit object from a global fit}

\item{is_married}{boolean}
}
\value{
NULL, but saves residual plots to output_dir
}
\description{
Plot residuals
}
\keyword{internal}
