% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subnational_comparison.R
\name{plot_subnational_comparison}
\alias{plot_subnational_comparison}
\title{Plot subnational comparison}
\usage{
plot_subnational_comparison(
  results,
  indicator_select,
  year_select,
  marital_status_select,
  ymin_select = 0,
  ymax_select = NA,
  arrange_point = TRUE
)
}
\arguments{
\item{results}{output from fit_fpem()}

\item{indicator_select}{choice out of c("contraceptive_use_modern", "unmet_need_modern",
"demand", "demand_satisfied_modern")}

\item{year_select}{year to plot}

\item{marital_status_select}{choice out of c("all", "married", "unmarried")}

\item{ymin_select}{default 0}

\item{ymax_select}{default NA}

\item{arrange_point}{logical, if TRUE arrange points by value of indicator_select}
}
\value{
a ggplot object
}
\description{
Plot subnational comparison
}
\keyword{internal}
