% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process survey data}
\usage{
process_data(dat, regions_dat, start_year = 1970, is_married = TRUE)
}
\arguments{
\item{dat}{Survey data}

\item{regions_dat}{File with region meta data}

\item{start_year}{Filter out data prior to this year}

\item{is_married}{Logical, select married women data?}
}
\value{
Processed survey data
}
\description{
Process survey data
}
