% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_all.R
\docType{data}
\name{regions_all}
\alias{regions_all}
\title{Region meta data}
\format{
A tibble, one row per country. Variables include
\describe{
  \item{division_numeric_code}{Divison numeric code}
  \item{name_country}{Name country}
  \item{cluster}{Name of cluster}
  \item{subcluster}{Name of subcluster}
  \item{is_unmarried_sexual_activity}{SA0 indicator}
}
}
\usage{
regions_all
}
\description{
Dataset with meta information on countries and their regional groupings.
}
\keyword{datasets}
