% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dir_related.R
\name{rename_output_folder}
\alias{rename_output_folder}
\title{Rename output folder and directory in fit object}
\usage{
rename_output_folder(indicator, run_step, old_folder_name, new_folder_name)
}
\arguments{
\item{indicator}{Character. Indicator name.}

\item{run_step}{Character. Run step - either 1a, 1b, local_national.}

\item{old_folder_name}{Character. The current folder name (inside `bayestransition_output`).}

\item{new_folder_name}{Character. The new name to rename the folder to.}
}
\value{
NULL. The function performs the renaming and updates the fit object in place.
}
\description{
Renames an existing folder in `bayestransition_output`and updates the `output_dir` in the model fit object.
}
\keyword{internal}
