% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{replacesomething_inblock}
\alias{replacesomething_inblock}
\title{Replace something in block}
\usage{
replacesomething_inblock(
  block,
  params = c("Epsilon", "Eta", "Betas", "Ptilde", "Omega", "Rho", "Tau"),
  prefix = "d_"
)
}
\arguments{
\item{block}{string with stan code block}

\item{params}{vector of parameter names to replace}

\item{prefix}{prefix to add to parameter names}
}
\value{
modified block
}
\description{
Replace something in block
}
\keyword{internal}
