% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{round_from_zero}
\alias{round_from_zero}
\title{Round from zero}
\usage{
round_from_zero(x)
}
\arguments{
\item{x}{proportion to round}
}
\value{
proportions less than 0.001 are rounded upwards to 0.001
}
\description{
Taken from UNPD/fpemlocal:
https://github.com/AlkemaLab/fpemlocal/blob/master/R/rounding_and_recalculation.R
}
\keyword{internal}
