% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emu_standata.R
\name{service_statistics_getstandata}
\alias{service_statistics_getstandata}
\title{Get Stan data for service statistics data}
\usage{
service_statistics_getstandata(
  service_statistic_df,
  time_index,
  hyper_param,
  geo_unit_index
)
}
\arguments{
\item{service_statistic_df}{emus filtered to pop (eg country)}

\item{time_index}{time_index used in model fit}

\item{hyper_param}{hyper parameters for EMUs to be used in model fit}

\item{geo_unit_index}{geo_unit_index used in model fit}
}
\value{
list with dat_emu (for plotting) and emu_list (to pass to stan)
}
\description{
This function is called from fit_fpem, to prepare service statistics data for use in model fitting
}
\keyword{internal}
