% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_name_country.R
\name{std_name_country}
\alias{std_name_country}
\title{Standardize country names}
\usage{
std_name_country(name_country)
}
\arguments{
\item{name_country}{country name}
}
\value{
Standardized country name
}
\description{
Standardize country names
}
\keyword{internal}
