% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{subpopulation_descriptions}
\alias{subpopulation_descriptions}
\title{Assign descriptions for subpopulations}
\usage{
subpopulation_descriptions(contraceptive_use)
}
\arguments{
\item{contraceptive_use}{data set with information on subgroups}
}
\value{
descriptions for subpopulations
}
\description{
Assign descriptions for subpopulations
}
\keyword{internal}
