% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemlocal_datacleaning.R
\name{subpopulation_labels}
\alias{subpopulation_labels}
\title{Assign labels for subpopulations}
\usage{
subpopulation_labels(contraceptive_use)
}
\arguments{
\item{contraceptive_use}{}
}
\value{
labels for subpopulations
}
\description{
Taken from UNPD/fpemlocal: see https://github.com/AlkemaLab/fpemlocal/blob/master/R/subpopulation_labels.R
}
\keyword{internal}
