% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates_samples_for_indicators.R
\name{summarize_fp_helper}
\alias{summarize_fp_helper}
\title{Summarize family planning indicators helper}
\usage{
summarize_fp_helper(samples)
}
\arguments{
\item{samples}{tibble with columns indicator and value, where
indicator includes (unmet_need_any, contraceptive_use_modern,
contraceptive_use_traditional)}
}
\value{
tibble with additional family planning indicators calculated
}
\description{
This function summarizes family planning indicators from samples
}
\keyword{internal}
