% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarize samples}
\usage{
summarize_samples(
  samples,
  percentiles = c(0.025, 0.05, 0.1, 0.5, 0.9, 0.95, 0.975)
)
}
\arguments{
\item{samples}{tibble with `value` column and other grouping columns}

\item{percentiles}{vector of percentiles to compute}
}
\value{
tibble with summary columns (mean and percentiles) for each group
}
\description{
Summarize samples
}
\keyword{internal}
