% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outliers.R
\name{update_nooutlier_in_global_fit}
\alias{update_nooutlier_in_global_fit}
\title{Use global fit 1a residuals to update assignment of outlier errors}
\usage{
update_nooutlier_in_global_fit(fit, is_married)
}
\arguments{
\item{fit}{global fit 1a}

\item{is_married}{logical}
}
\value{
data frame with updated nooutlier column.
}
\description{
In this function, we calculate residuals from global fit 1a to
update the `nooutlier` column for use in further model fitting.
In this function, `nooutlier` is set to 0 for observations that are outlying in 1a,
here defined as observations with absolute standardized residuals above the 90th percentile.
By setting `nooutlier` to zero, outlier errors are assigned.
(Naming of the column can use an update!)
}
