% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write stan models}
\usage{
write_model(
  add_aggregates = FALSE,
  add_trad = TRUE,
  all_women = FALSE,
  add_emu = FALSE
)
}
\arguments{
\item{add_aggregates}{Add national aggregates? makes sense only for subnational all-region run for 1 country}

\item{add_trad}{Include trad use estimation and data?}

\item{all_women}{Estimate for all_women or by marital group?}

\item{add_emu}{Include EMU estimation and data?}
}
\value{
writes a stan file to the stan directory. The model files are named
based on the options chosen.
}
\description{
Write stan models
}
